Moving network calls into their own folder
