--- tickspot/__init__.py
@@ -1,3 +1,0 @@
-from .manual_settings import *
-from .network import *
-from .tickspot import *
